feat Add Authorization to Comment

+++ api/controller/CommentController.cs
@@ -32,13 +32,17 @@ public class CommentController : ControllerBase
     [HttpGet("{id}")]
     [Authorize]
     public async Task<IActionResult> GetById([FromRoute] int id){
-        var commnet = await _commentRepo.GetByIdAsync(id);
-        Console.WriteLine(commnet);
-     
-        if(commnet is null){
+        var comment = await _commentRepo.GetByIdAsync(id);
+        string username = User.GetUsername();
+        var appUser = await _userManager.FindByNameAsync(username);
+        
+        if(comment is null){
             return NotFound();
         }
-        return Ok(commnet.ToCommentDto());
+        if(comment.AppUserId != appUser?.Id){
+            return Forbid();
+        }
+        return Ok(comment.ToCommentDto());
     }
     [HttpPost("{stockId}")]
     [Authorize]
@@ -46,14 +50,14 @@ public class CommentController : ControllerBase
         if(!ModelState.IsValid){
             return BadRequest(ModelState);
         }
+
         if(!await _stockRepo.StockExists(stockId)){
             return BadRequest("Stock does not exist");
         }
 
         var username = User.GetUsername();
         var appUser = await _userManager.FindByNameAsync(username);
-
-
+        Console.WriteLine(appUser);
         var commentModel = commentDto.ToCommnetFromCreate(stockId);
         commentModel.AppUserId = appUser.Id;
         await _commentRepo.CreateAsync(commentModel);
@@ -61,27 +65,36 @@ public class CommentController : ControllerBase
     }
     [HttpPut("{id}")]
     [Authorize]
-
     public async Task<IActionResult> Update([FromRoute] int id, UpdateCommentRequestDto updateCommentRequestDto){
         if(!ModelState.IsValid){
             return BadRequest(ModelState);
         }
-        string username = User.GetUsername();
+        string username =  User.GetUsername();
         var appUser = await _userManager.FindByNameAsync(username);
-
-        var commnet = await _commentRepo.UpdateAsync(id, updateCommentRequestDto.ToCommentFromUpdate());
-        if(commnet is null){
+        var comment = await _commentRepo.GetByIdAsync(id);
+        if(comment is null){
             return NotFound();
         }
-        return Ok(commnet.ToCommentDto());
+        if(comment?.AppUserId != appUser?.Id){
+            return Forbid();
+        }
+        var updatedComment = await _commentRepo.UpdateAsync(id, updateCommentRequestDto.ToCommentFromUpdate());
+        
+        return Ok(updatedComment?.ToCommentDto());
     }
     
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteAsync([FromRoute] int id){
-        var commnet = await _commentRepo.DeleteAsync(id);
-        if(commnet is null){
-            return NotFound();
+        var comment = await _commentRepo.GetByIdAsync(id);
+        if(comment is null){
+            NotFound();
         }
+        string username = User.GetUsername();
+        var appUser = await _userManager.FindByNameAsync(username);
+        if(comment?.AppUserId != appUser?.Id){
+            return Forbid();
+        }
+        await _commentRepo.DeleteAsync(id);
         return NoContent();
     }
 }
